requirements.txt
Flask
requests
gunicorn
beautifulsoup4
google-cloud-storage
yfinance
pandas